+++ back-end/src/main/java/io/thescenery/shopAssist/user/mapper/UserMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Param;
 @Mapper
 public interface UserMapper {
 
-  User getUserById(@Param("id") int id);
+  User getUserById(@Param("id") long id);
 
   User getUserByEmail(@Param("email") String email);
 }
