+++ back-end/src/main/java/io/thescenery/shopAssist/user/service/impl/UserServiceImpl.java
@@ -29,7 +29,7 @@ public class UserServiceImpl implements IUserService {
   }
 
   @Override
-  public UserInfoVo getUserById(int id) {
+  public UserInfoVo getUserById(long id) {
     return userEntityToUserVo(userMapper.getUserById(id));
   }
